Fix pytest failures when the test env has UTC as local timezone
Some of the Linux commands that stage_check may execute on a remote node
output local datetime, i.e. 'Wed 2020-01-01 12:00:00 EDT' instead of UTC.

In order to perform local comparisons, the datetimes are converted to
UTC on the local node.  However the local node may not recognize the
timezone name of the remote node.

An admittedly less than perfect workaround is to temporarily change the
timezone name of the local node to the timezone name of the remote node,
and then convert the remote datetime to UTC (This would be less of a
problem if the Linux commands in questiom output time deltas from GMT
instead of a timezone name).